+++ src/cmd/go/internal/work/cover.go
@@ -10,11 +10,11 @@ import (
 	"cmd/go/internal/base"
 	"cmd/go/internal/cfg"
 	"cmd/go/internal/str"
+	"cmd/internal/cov/covcmd"
 	"context"
 	"encoding/json"
 	"fmt"
 	"internal/coverage"
-	"internal/coverage/covcmd"
 	"io"
 	"os"
 	"path/filepath"
