+++ src/cmd/internal/cov/covcmd/cmddefs.go
